refactor([issue-1148]): split taskPromptDefaults.js into a directory module with a re-exporting barrel
2606 lines of prompt prose around 3 runtime exports now live in
taskPromptDefaults/prompts.js, versions.js, and previousDefaults.js, with
taskPromptDefaults.js left as a barrel so all existing imports keep working.
The split was done by line-range extraction and verified byte-identical
against the pre-split exports; taskPromptDefaults.test.js pins every prompt
body and version against a hash snapshot (PORTOS_API_URL normalized to a
placeholder so the test is env-stable), making future drift in the
cross-install prompt-upgrade contract loud.

Refs #1148
